requirements.txt
fastapi==0.115.0
pydantic[email]==2.9.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
Pillow==10.4.0
pytesseract==0.3.13
google-generativeai==0.8.3
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1
cryptography==43.0.3
